Migrate addUserToGuild and removeUserFromGuild to Lilac
